requirements.txt
redis>=4.5.0
sqlalchemy>=2.0.0
llama-cpp-python>=0.2.0
pyyaml>=6.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
flake8>=6.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
python-multipart>=0.0.6
httpx>=0.25.0
supabase>=1.0.3
postgrest>=0.10.8 